[Handlers] Move request/response structs out
Move the response/request structs out into the structs
package.

This will allow us to more easily see at a glance what the
various response/request structs are, and hopefully
we can actually condense them down to just a few.

+++ sept-2018/structs/node.go
@@ -3,7 +3,7 @@ package structs
 type Node struct {
 	GridEntity
 	ClaimedBy string
-	Value     uint64
+	Value     int
 }
 
 type NodeStatus struct {
